# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /setup/rule-profile:
    get:
      operationId: getRuleProfile
      tags: [setup]
      summary: Get the org rule profile
      responses:
        "200":
          description: Rule profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RuleProfile"
    patch:
      operationId: updateRuleProfile
      tags: [setup]
      summary: Update the org rule profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RuleProfileUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RuleProfile"
...
  /maintenance/work-orders:
    post:
      operationId: createWorkOrder
      tags: [maintenance]
      summary: Create a work order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkOrderInput"
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkOrderItem"
  /maintenance/work-orders/{id}:
    patch:
      operationId: updateWorkOrder
      tags: [maintenance]
      summary: Update a work order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkOrderUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkOrderItem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorMessage"
...
  /setup/alert-rules/{kind}:
    patch:
      operationId: updateAlertRule
      tags: [setup]
      summary: Update (upsert) one alert rule by kind
      parameters:
        - name: kind
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlertRuleUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertRule"
  /integrations/api-keys:
    get:
      operationId: getApiKeys
      tags: [integrations]
      summary: List API keys
      responses:
        "200":
          description: API keys
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApiKeyItem"
    post:
      operationId: createApiKey
      tags: [integrations]
      summary: Create an API key (plaintext returned once)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApiKeyInput"
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiKeyCreated"
  /integrations/api-keys/{id}:
    delete:
      operationId: revokeApiKey
      tags: [integrations]
      summary: Revoke an API key
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Revoked
  /integrations/webhooks:
    get:
      operationId: getWebhooks
      tags: [integrations]
      summary: List webhooks
      responses:
        "200":
          description: Webhooks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WebhookItem"
    post:
      operationId: createWebhook
      tags: [integrations]
      summary: Create a webhook subscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebhookInput"
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookItem"
  /integrations/webhooks/{id}:
    delete:
      operationId: deleteWebhook
      tags: [integrations]
      summary: Delete a webhook
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /ingest/telemetry:
    post:
      operationId: ingestTelemetry
      tags: [integrations]
      summary: Ingest a device telemetry batch (authenticated by API key)
      description: >
        External device gateways POST batches of telemetry here using an API key
        in the `x-api-key` header. The org is resolved from the key, not a session.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TelemetryIngest"
      responses:
        "200":
          description: Accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestReceipt"
        "401":
          description: Invalid or missing API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorMessage"
components:
  schemas:
...
    ApiErrorMessage:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    RuleProfile:
      type: object
      required:
        - id
        - name
        - contMins
        - dailyMins
        - weeklyMins
        - dailyRestMins
        - breakMins
      properties:
        id:
          type: string
        name:
          type: string
        contMins:
          type: integer
        dailyMins:
          type: integer
        weeklyMins:
          type: integer
        dailyRestMins:
          type: integer
        breakMins:
          type: integer
    RuleProfileUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        contMins:
          type: integer
        dailyMins:
          type: integer
        weeklyMins:
          type: integer
        dailyRestMins:
          type: integer
        breakMins:
          type: integer
...
    WorkOrderItem:
      type: object
      required: [id, vehicleId, vehicleReg, title, status, priority, createdAt]
      properties:
        id:
          type: string
        vehicleId:
          type: string
        vehicleReg:
          type: string
        title:
          type: string
        detail:
          type: ["string", "null"]
        status:
          type: string
        priority:
          type: string
        dueOn:
          type: ["string", "null"]
        dueKm:
          type: ["integer", "null"]
        costEstimate:
          type: ["integer", "null"]
        createdAt:
          type: string
        completedOn:
          type: ["string", "null"]
    WorkOrderInput:
      type: object
      required: [vehicleId, title]
      properties:
        vehicleId:
          type: string
          minLength: 1
        title:
          type: string
          minLength: 1
        detail:
          type: string
        priority:
          type: string
          enum: [LOW, MEDIUM, HIGH]
        dueOn:
          type: string
        dueKm:
          type: integer
        costEstimate:
          type: integer
    WorkOrderUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        detail:
          type: string
        status:
          type: string
          enum: [OPEN, IN_PROGRESS, DONE, CANCELLED]
        priority:
          type: string
          enum: [LOW, MEDIUM, HIGH]
        dueOn:
          type: string
        dueKm:
          type: integer
        costEstimate:
          type: integer
...
    AlertRule:
      type: object
      required: [kind, enabled, severity, notifyEmail, notifySms, notifyPush]
      properties:
        kind:
          type: string
        enabled:
          type: boolean
        threshold:
          type: ["integer", "null"]
        severity:
          type: string
        notifyEmail:
          type: boolean
        notifySms:
          type: boolean
        notifyPush:
          type: boolean
    AlertRuleUpdate:
      type: object
      properties:
        enabled:
          type: boolean
        threshold:
          type: ["integer", "null"]
        severity:
          type: string
          enum: [LOW, MEDIUM, HIGH]
        notifyEmail:
          type: boolean
        notifySms:
          type: boolean
        notifyPush:
          type: boolean
    ApiKeyItem:
      type: object
      required: [id, name, prefix, scopes, createdAt, revoked]
      properties:
        id:
          type: string
        name:
          type: string
        prefix:
          type: string
        scopes:
          type: array
          items:
            type: string
        lastUsedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        revoked:
          type: boolean
    ApiKeyInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        scopes:
          type: array
          items:
            type: string
    ApiKeyCreated:
      type: object
      required: [id, name, prefix, key, scopes, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        prefix:
          type: string
        key:
          type: string
          description: The plaintext key, shown only once at creation.
        scopes:
          type: array
          items:
            type: string
        createdAt:
          type: string
    WebhookItem:
      type: object
      required: [id, url, events, active, createdAt]
      properties:
        id:
          type: string
        url:
          type: string
        events:
          type: array
          items:
            type: string
        active:
          type: boolean
        createdAt:
          type: string
        lastDeliveryAt:
          type: ["string", "null"]
        lastStatus:
          type: ["integer", "null"]
    WebhookInput:
      type: object
      required: [url, events]
      properties:
        url:
          type: string
          minLength: 1
        events:
          type: array
          items:
            type: string
    TelemetryPingInput:
      type: object
      required: [vehicleReg, lat, lng, speedKph, recordedAt]
      properties:
        vehicleReg:
          type: string
        lat:
          type: number
        lng:
          type: number
        speedKph:
          type: integer
        placeLabel:
          type: string
        recordedAt:
          type: string
    TelemetryIngest:
      type: object
      required: [pings]
      properties:
        pings:
          type: array
          items:
            $ref: "#/components/schemas/TelemetryPingInput"
    IngestReceipt:
      type: object
      required: [accepted, rejected]
      properties:
        accepted:
          type: integer
        rejected:
          type: integer