# lib/api-zod/src/generated/types/apiKeyItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ApiKeyItem {
  id: string;
  name: string;
  prefix: string;
  scopes: string[];
  /** @nullable */
  lastUsedAt?: string | null;
  createdAt: string;
  revoked: boolean;
}