# lib/api-zod/src/generated/types/telemetryPingInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TelemetryPingInput {
  vehicleReg: string;
  lat: number;
  lng: number;
  speedKph: number;
  placeLabel?: string;
  recordedAt: string;
}