# lib/api-zod/src/generated/types/webhookInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WebhookInput {
  /** @minLength 1 */
  url: string;
  events: string[];
}