# lib/api-zod/src/generated/types/workOrderUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { WorkOrderUpdatePriority } from './workOrderUpdatePriority';
import type { WorkOrderUpdateStatus } from './workOrderUpdateStatus';

export interface WorkOrderUpdate {
  /** @minLength 1 */
  title?: string;
  detail?: string;
  status?: WorkOrderUpdateStatus;
  priority?: WorkOrderUpdatePriority;
  dueOn?: string;
  dueKm?: number;
  costEstimate?: number;
}